+++ examples/arrow/arrow-instancing/arrow-instanced-mesh-renderer.ts
@@ -52,7 +52,7 @@ struct FragmentInputs {
   @builtin(position) Position : vec4<f32>,
   @location(0) normal : vec3<f32>,
   @location(1) color : vec4<f32>,
-  @interpolate(flat)
+  @interpolate(flat, either)
   @location(2) objectIndex : i32,
 }